requirements.txt
Django==5.0.6
djangorestframework==3.15.1
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.2
pytest-django==4.8.0
